refactor workbench performance routes
